implicitly inline things, separate pins from module headers
